+++ demo/src/main/resources/com/romei/main.fxml
@@ -24,8 +24,8 @@
       <Button fx:id="updProductsView" text="Editar Productos" onAction="#loadUpdProductsView"/>
       <Button fx:id="addUsersView" text="Agregar Usuarios" onAction="#loadAddUsersView"/>
       <Button fx:id="updUsersView" text="Editar Usuarios" onAction="#loadUpdUsersView"/>
-      <Button fx:id="addSuppliersView" text="Agregar Usuarios" onAction="#loadAddSuppliersView"/>
-      <Button fx:id="updSuppliersView" text="Editar Usuarios" onAction="#loadUpdSuppliersView"/>
+      <Button fx:id="addSuppliersView" text="Agregar Proveedores" onAction="#loadAddSuppliersView"/>
+      <Button fx:id="updSuppliersView" text="Editar Proveedores" onAction="#loadUpdSuppliersView"/>
       <Button fx:id="logout" text="Cerrar sesión" onAction="#runLogout"/>
       <Button fx:id="exit" text="Exit" onAction="#runExit"/>
       </VBox>
